server/requirements.txt

Flask 
Flask-SQLAlchemy 
Flask-Migrate 
Flask-RESTful 
Flask-CORS
Flask-Bcrypt 
python-dotenv 
Authlib 
Flask-Login 
requests
